TINY-13548: rename origin predicates file
