+++ modules/oxide-components/src/main/ts/components/draggable/internals/OriginPredicates.ts
